Received more packets than expected - there must be a duplicate retransmission that is happening.

+++ examples/common/thread_border_router/include/handler.h
@@ -15,8 +15,8 @@ typedef struct PacketLossStats
 {
   struct timeval start;
   uint32_t nextSeqNumExpected;
-  uint64_t packetsReceived;
-  uint64_t packetsExpected;
+  uint32_t packetsReceived;
+  uint32_t packetsExpected;
   PacketLossExpState state;
 }
 PacketLossStats;
